Finish up informer implementation
Still somewhat inefficient, because we reconcile separately for each
node at start-up.
